backend/requirements.txt
# Backend Requirements - Rick Sanchez Chatbot
# Python 3.11+

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# OpenAI API
openai==1.10.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3

# HTTP Requests (para extensões futuras)
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS Middleware (já incluído no FastAPI, mas explícito)
# fastapi já inclui starlette

# Code Quality (opcional, para desenvolvimento)
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0
